Added Response Type attributes to AccountController methods.

+++ ProjectForFarmers.WebApi/Controllers/AccountController.cs
@@ -1,9 +1,14 @@
 ﻿using FarmersMarketplace.Application.DataTransferObjects;
 using FarmersMarketplace.Application.DataTransferObjects.Account;
+using FarmersMarketplace.Application.Exceptions;
+using FarmersMarketplace.Application.Helpers;
 using FarmersMarketplace.Application.Services.Business;
+using FarmersMarketplace.Application.ViewModels.Account;
+using FarmersMarketplace.Application.ViewModels.Auth;
 using FarmersMarketplace.Domain;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
+using InvalidDataException = FarmersMarketplace.Application.Exceptions.InvalidDataException;
 
 namespace FarmersMarketplace.WebApi.Controllers
 {
@@ -20,15 +25,15 @@ namespace FarmersMarketplace.WebApi.Controllers
         }
 
         [HttpGet("{accountId}")]
+        [ProducesResponseType(typeof(CustomerVm), 200)]
         public async Task<IActionResult> GetCustomer([FromRoute] Guid accountId)
         {
-            var customer = await AccountService.GetCustomer(accountId);
-            if (customer == null)
-                return NotFound();
-            return Ok(customer);
+            var vm = await AccountService.GetCustomer(accountId);
+            return Ok(vm);
         }
 
         [HttpPut]
+        [ProducesResponseType(204)]
         public async Task<IActionResult> UpdateCustomer([FromForm] UpdateCustomerDto dto)
         {
             await AccountService.UpdateCustomer(dto, AccountId);
@@ -36,6 +41,7 @@ namespace FarmersMarketplace.WebApi.Controllers
         }
 
         [HttpPut]
+        [ProducesResponseType(204)]
         public async Task<IActionResult> UpdateCustomerPaymentData([FromBody] CustomerPaymentDataDto dto)
         {
             await AccountService.UpdateCustomerPaymentData(dto, AccountId);
@@ -43,13 +49,15 @@ namespace FarmersMarketplace.WebApi.Controllers
         }
 
         [HttpGet("{accountId}")]
+        [ProducesResponseType(typeof(SellerVm), 200)]
         public async Task<IActionResult> GetSeller([FromRoute] Guid accountId)
         {
-            var seller = await AccountService.GetSeller(accountId);
-            return Ok(seller);
+            var vm = await AccountService.GetSeller(accountId);
+            return Ok(vm);
         }
 
         [HttpPut]
+        [ProducesResponseType(204)]
         public async Task<IActionResult> UpdateSeller([FromForm] UpdateSellerDto dto)
         {
             await AccountService.UpdateSeller(dto, AccountId);
@@ -57,6 +65,7 @@ namespace FarmersMarketplace.WebApi.Controllers
         }
 
         [HttpPut]
+        [ProducesResponseType(204)]
         public async Task<IActionResult> UpdateSellerCategoriesAndSubcategories([FromBody] SellerCategoriesAndSubcategoriesDto dto)
         {
             await AccountService.UpdateSellerCategoriesAndSubcategories(dto, AccountId);
@@ -64,13 +73,15 @@ namespace FarmersMarketplace.WebApi.Controllers
         }
 
         [HttpGet("{accountId}")]
+        [ProducesResponseType(typeof(FarmerVm), 200)]
         public async Task<IActionResult> GetFarmer([FromRoute] Guid accountId)
         {
-            var farmer = await AccountService.GetFarmer(accountId);
-            return Ok(farmer);
+            var vm = await AccountService.GetFarmer(accountId);
+            return Ok(vm);
         }
 
         [HttpPut]
+        [ProducesResponseType(204)]
         public async Task<IActionResult> UpdateFarmer([FromForm] UpdateFarmerDto dto)
         {
             await AccountService.UpdateFarmer(dto, AccountId);
@@ -78,15 +89,27 @@ namespace FarmersMarketplace.WebApi.Controllers
         }
 
         [HttpPut("{producer}")]
+        [ProducesResponseType(204)]
         public async Task<IActionResult> UpdateProducerPaymentData([FromBody] ProducerPaymentDataDto dto, [FromRoute] Role producer)
         {
             await AccountService.UpdateProducerPaymentData(dto, AccountId, producer);
             return NoContent();
         }
 
-        [HttpDelete("{role}")]
-        public async Task<IActionResult> DeleteAccount([FromRoute] Role role)
+        [HttpDelete]
+        [ProducesResponseType(204)]
+        public async Task<IActionResult> DeleteAccount()
         {
+            var roleStr = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (string.IsNullOrEmpty(roleStr) || !Enum.TryParse<Role>(roleStr, out var role))
+            {
+                string message = $"Failed to retrieve or parse role claim from JWT.";
+                string userFacingMessage = CultureHelper.Exception("IncorrectRole");
+
+                throw new InvalidDataException(message, userFacingMessage);
+            }
+
             await AccountService.DeleteAccount(role, AccountId);
             return NoContent();
         }
